Move code to pkg/querier/correctness.
