+++ pkg/querier/correctness/time_flag.go
@@ -1,4 +1,4 @@
-package test
+package correctness
 
 import (
 	"time"
